Fix some compiler warnings

+++ drivers/amlogic/mmc/amlsd.h
@@ -73,10 +73,10 @@ void aml_sd_uart_detect_clr (struct amlsd_platform* pdata);
 		prop = 0;													\
 	    if (DEBUG_SD_OF) {                                          \
 	        printk("get property:%25s, value:0x%08x\n",    			\
-	            prop_name, value);                           		\
+	            prop_name, (unsigned int)value);                           		\
 	    }															\
     }
-	
+
 #define SD_PARSE_GPIO_NUM_PROP(node, prop_name, str, gpio_pin)		\
 	if(!of_property_read_string(node, prop_name, &str)) { 			\
 		gpio_pin = amlogic_gpio_name_map_num(str);					\
@@ -85,7 +85,7 @@ void aml_sd_uart_detect_clr (struct amlsd_platform* pdata);
 	            prop_name, str);    								\
 		}															\
 	}
-	
+
 #define SD_PARSE_STRING_PROP(node, prop_name, str, prop)      		\
 	if (!of_property_read_string(node, prop_name, &str)){ 			\
 		strcpy(prop, str);											\
